remove dcf and qdata
