+++ src/mrpro/algorithms/reconstruction/RegularizedIterativeSENSEReconstruction.py
@@ -10,10 +10,10 @@ from mrpro.algorithms.optimizers.cg import cg
 from mrpro.algorithms.prewhiten_kspace import prewhiten_kspace
 from mrpro.algorithms.reconstruction.DirectReconstruction import DirectReconstruction
 from mrpro.data.CsmData import CsmData
-from mrpro.data.DcfData import DcfData
 from mrpro.data.IData import IData
 from mrpro.data.KData import KData
 from mrpro.data.KNoise import KNoise
+from mrpro.operators.DensityCompensationOp import DensityCompensationOp
 from mrpro.operators.IdentityOp import IdentityOp
 from mrpro.operators.LinearOperator import LinearOperator
 from mrpro.utils import unsqueeze_right
@@ -48,7 +48,7 @@ class RegularizedIterativeSENSEReconstruction(DirectReconstruction):
         fourier_op: LinearOperator | None = None,
         csm: Callable | CsmData | None = CsmData.from_idata_walsh,
         noise: KNoise | None = None,
-        dcf: DcfData | None = None,
+        dcf_op: DensityCompensationOp | None = None,
         *,
         n_iterations: int = 5,
         regularization_data: float | torch.Tensor = 0.0,
@@ -77,10 +77,10 @@ class RegularizedIterativeSENSEReconstruction(DirectReconstruction):
             or `~mrpro.data.CsmData.from_idata_inati`.
         noise
             Noise used for prewhitening. If `None`, no prewhitening is performed
-        dcf
-            K-space sampling density compensation. If `None`, set up based on `kdata`.
-            Used to obtain a the starting point for the CG algorithm as the scaled density compensated direct
-            reconstruction [FESSLER2010]_.
+        dcf_op
+            Instance of the `~mrpro.operators.DensityCompensationOp` to compensate for the k-space sampling density.
+            If `None`, set up based on `kdata`. Only used to obtain a good starting point for the CG algorithm as
+            by using the scaled density compensated direct reconstruction [FESSLER2010]_.
         n_iterations
             Number of CG iterations
         regularization_data
@@ -101,7 +101,7 @@ class RegularizedIterativeSENSEReconstruction(DirectReconstruction):
         `ValueError`
             If the `kdata` and `fourier_op` are `None` or if `csm` is a `Callable` but `kdata` is None.
         """
-        super().__init__(kdata, fourier_op, csm, noise, dcf)
+        super().__init__(kdata, fourier_op, csm, noise, dcf_op)
         self.n_iterations = n_iterations
         self.regularization_data = torch.as_tensor(regularization_data)
         self.regularization_weight = torch.as_tensor(regularization_weight)
@@ -135,13 +135,12 @@ class RegularizedIterativeSENSEReconstruction(DirectReconstruction):
                 right_hand_side + self.regularization_weight * self.regularization_op.H(self.regularization_data)[0]
             )
 
-        if self.dcf is not None:
+        if self.dcf_op is not None:
             # The DCF is used to obtain a good starting point for the CG algorithm.
             # This is equivalten to running the CG algorithm with H = A^H DCF A and b = A^H DCF y
             # for a single iteration.
-            dcf_op = self.dcf.as_operator()
-            (u,) = (acquisition_model.H @ dcf_op)(kdata.data)
-            (v,) = (acquisition_model.H @ dcf_op @ acquisition_model)(u)
+            (u,) = (acquisition_model.H @ self.dcf_op)(kdata.data)
+            (v,) = (acquisition_model.H @ self.dcf_op @ acquisition_model)(u)
             u_flat = u.flatten(start_dim=-3)
             v_flat = v.flatten(start_dim=-3)
             initial_value = (
