+++ src/mrpro/data/__init__.py
@@ -4,7 +4,6 @@ from mrpro.data import enums, traj_calculators, acq_filters
 from mrpro.data.AcqInfo import AcqIdx, AcqInfo
 from mrpro.data.CsmData import CsmData
 from mrpro.data.Dataclass import Dataclass, InconsistentDeviceError
-from mrpro.data.DcfData import DcfData
 from mrpro.data.EncodingLimits import EncodingLimits, Limits
 from mrpro.data.IData import IData
 from mrpro.data.IHeader import IHeader
@@ -12,8 +11,6 @@ from mrpro.data.KData import KData
 from mrpro.data.KHeader import KHeader
 from mrpro.data.KNoise import KNoise
 from mrpro.data.KTrajectory import KTrajectory
-from mrpro.data.QData import QData
-from mrpro.data.QHeader import QHeader
 from mrpro.data.Rotation import Rotation
 from mrpro.data.SpatialDimension import SpatialDimension
 
@@ -22,7 +19,6 @@ __all__ = [
     "AcqInfo",
     "CsmData",
     "Dataclass",
-    "DcfData",
     "EncodingLimits",
     "IData",
     "IHeader",
@@ -32,8 +28,6 @@ __all__ = [
     "KNoise",
     "KTrajectory",
     "Limits",
-    "QData",
-    "QHeader",
     "Rotation",
     "SpatialDimension",
     "acq_filters",
